--- hooks/src/lib/hooks/index.ts
@@ -1,6 +1,0 @@
-export { default as useCardIssuer } from './useInputCardIssuer';
-export { default as useCardNumber } from './useInputCardNumber';
-export { default as useCVCNumber } from './useInputCVCNumber';
-export { default as useExpiryDate } from './useInputExpiryDate';
-export { default as useOwnerName } from './useInputOwnerName';
-export { default as usePasswordPrefix } from './useInputPasswordPrefix';
